Quick_edu/requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
billiard==4.2.1
cashfree-pg==4.3.10
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.3
daphne==4.1.2
defusedxml==0.7.1
Django==5.2
django-allauth==65.8.0
django-appconf==1.1.0
django-celery-beat==2.8.0
django-celery-email==3.0.0
django-countries==7.6.1
django-crontab==0.7.1
django-extensions==4.1
django-minio-backend==3.8.0
django-minio-storage==0.5.8
django-phonenumber-field==8.1.0
django-phonenumbers==1.0.1
django-timezone-field==7.1
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
Faker==37.1.0
frozenlist==1.6.0
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
minio==7.2.15
msgpack==1.1.0
multidict==6.4.3
mysqlclient==2.2.7
numpy==2.2.5
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
paypalrestsdk==1.13.3
phonenumbers==9.0.3
pillow==11.2.1
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==1.10.22
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
pyuca==1.2
razorpay==1.4.2
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
sentry-sdk==1.32.0
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
text-unidecode==1.3
tomli==2.2.1
twilio==9.5.2
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.0.7
vine==5.1.0
vonage==4.4.1
vonage-account==1.1.1
vonage-application==2.0.1
vonage-http-client==1.5.1
vonage-jwt==1.1.5
vonage-messages==1.4.0
vonage-network-auth==1.0.2
vonage-network-number-verification==1.0.2
vonage-network-sim-swap==1.1.2
vonage-number-insight==1.0.6
vonage-numbers==1.0.4
vonage-sms==1.1.5
vonage-subaccounts==1.0.4
vonage-users==1.2.1
vonage-utils==1.1.4
vonage-verify==2.1.0
vonage-verify-legacy==1.0.1
vonage-video==1.2.0
vonage-voice==1.4.0
wcwidth==0.2.13
XlsxWriter==3.2.3
yarl==1.20.0
zope.interface==7.2
